Extract codegen.py: the shared code-building machinery
Step 3. Every helper tool does the same thing -- take caller parameters,
build a Sage snippet around them, run it -- and that machinery is now its
own module: the prelude, the literal encoding, the validation gates, the
numeric guards and _evaluate_structured.

Worth reading as a unit, because the gates are what stands between a tool
parameter and arbitrary execution: generated code runs under
trusted_policy(), which re-permits sage_eval, so a string interpolated
without passing _encode_literal, _validated_expression or
_validated_identifier is item 18 again. The module docstring says so.

server.py imports the names it still uses, so
`from sagemath_mcp.server import _validated_expression` keeps working for
the two tests that do it.

357 unit tests pass and the tool inventory is unchanged.
